Move devd address parsing into separate function
The debugfs command parses multiple devd addresses into a list. Split that code
into its own function so it can be reused in the manifest server.
