+++ shared/mount.c
@@ -16,7 +16,6 @@
 #include <netinet/tcp.h>
 
 #include "shared/lk/kernel.h"
-#include "shared/lk/limits.h"
 #include "shared/lk/list.h"
 #include "shared/lk/types.h"
 
@@ -52,47 +51,23 @@ static struct option_more mount_moreopts[] = {
 static int parse_mount_opt(int c, char *str, void *arg)
 {
 	struct mount_options *opts = arg;
-	struct ngnfs_manifest_addr_head *ahead;
 	int ret = -EINVAL;
 
 	switch(c) {
 	case 'd':
-		if (opts->nr_addrs == U8_MAX) {
-			log("too many -d addresses specified, exceeded limit of %u", U8_MAX);
-			ret = -EINVAL;
-			goto out;
-		}
-
-		ahead = malloc(sizeof(struct ngnfs_manifest_addr_head));
-		if (!ahead) {
-			ret = -ENOMEM;
-			goto out;
-		}
-
-		ret = parse_ipv4_addr_port(&ahead->addr, str);
-		if (ret < 0) {
-			log("error parsing -d address");
-			goto out;
-		}
-
-		list_add_tail(&ahead->head, &opts->addr_list);
-		opts->nr_addrs++;
+		ret = ngnfs_manifest_append_addr(&opts->nr_addrs, &opts->addr_list, str);
 		break;
 	case 't':
 		ret = strdup_nerr(&opts->trace_path, str);
 		break;
 	}
 
-	ret = 0;
-out:
 	return ret;
 }
 
 int ngnfs_mount(struct ngnfs_fs_info *nfi, int argc, char **argv)
 {
 	struct mount_options opts = { .addr_list = LIST_HEAD_INIT(opts.addr_list), };
-	struct ngnfs_manifest_addr_head *ahead;
-	struct ngnfs_manifest_addr_head *tmp;
 	int ret;
 
 	ret = getopt_long_more(argc, argv, mount_moreopts, ARRAY_SIZE(mount_moreopts),
@@ -114,10 +89,7 @@ out:
 	if (ret < 0)
 		ngnfs_unmount(nfi);
 
-	list_for_each_entry_safe(ahead, tmp, &opts.addr_list, head) {
-		list_del_init(&ahead->head);
-		free(ahead);
-	}
+	ngnfs_manifest_free_addrs(&opts.addr_list);
 
 	return ret;
 }
